refactor: 로또 컨트롤러 생성
